requirements.txt
discord.py
flask
yt-dlp
ffmpeg-python
py-webrtcvad